% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_template}
\alias{default_template}
\title{Default schema template}
\format{
A named character vector with fields:
\describe{
  \item{id}{SQLite type.}
  \item{numeric_feature}{SQLite type.}
  \item{text_feature}{SQLite type.}
  \item{logical_feature}{SQLite type.}
}
}
\usage{
default_template
}
\description{
Named character vector mapping column names to SQLite types.
}
\keyword{datasets}
