% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_template}
\alias{example_template}
\title{Example schema template}
\format{
A named character vector of length 3.
}
\usage{
example_template
}
\description{
Named character vector mapping column names to SQLite types.
}
\keyword{datasets}
