% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_sentences}
\alias{random_sentences}
\title{Random sentences}
\format{
A data frame with 10 rows and 2 columns:
\describe{
  \item{id}{Character identifier.}
  \item{text}{Sentence text.}
}
}
\usage{
random_sentences
}
\description{
Toy text data for examples/tests.
}
\keyword{datasets}
